ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.32.2
python-dotenv==1.0.1
llama-cpp-python==0.3.4
huggingface-hub==0.23.4
